refactor: LottoGame에서 로또 구입과 통계 출력부분을 별개의 서비스로 변경
