+++ src/main/java/lotto/Main.java
@@ -1,7 +1,9 @@
 package lotto;
 
 import lotto.service.LottoService;
+import lotto.service.StatisticPrinter;
 import lotto.view.ConsoleView;
+import lotto.view.View;
 import lotto.view.inputview.ConsoleInputView;
 import lotto.view.inputview.InputView;
 import lotto.view.resultview.ConsoleResultView;
@@ -14,7 +16,11 @@ public class Main {
     }
 
     private static LottoGame lottoGame() {
-        return new LottoGame(new ConsoleView(inputView(), resultView()), new LottoService());
+        return new LottoGame(new LottoService(view()), new StatisticPrinter(view()));
+    }
+
+    private static View view() {
+        return new ConsoleView(inputView(), resultView());
     }
 
     private static InputView inputView() {
